third_party/blink/web_tests/fast/css/lang-mapped-to-webkit-locale-dynamic-expected.txt

This tests mapping of lang/xml:lang to -webkit-locale with dynamic attribute changes. WebKit bug 76364

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS localeOfNode('x1') is "\"ko\""
PASS localeOfNode('x2') is "\"ko\""
PASS localeOfNode('x3') is "\"ar\""
PASS localeOfNode('x1') is "\"ja\""
PASS localeOfNode('x2') is "\"ja\""
PASS localeOfNode('x3') is "\"ar\""
PASS localeOfNode('x3') is "auto"
PASS localeOfNode('x3') is "\"ja\""
PASS localeOfNode('x2') is "\"zh\""
PASS localeOfNode('x2') is "\"ja\""
PASS localeOfNode('x2') is "\"ja\""
PASS successfullyParsed is true

TEST COMPLETE
